pyproject.toml
[tool.poetry]
name = "pubmed-fetcher"
version = "0.1.0"
description = "PubMed paper fetcher with non-academic affiliation filtering"
authors = ["Sunkeerth  sunkeerth123@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
biopython = "^1.80"
pandas = "^1.3.0"
tqdm = "^4.66.0"

[tool.poetry.scripts]
get-papers-list = "pubmed_fetcher.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"